unify prompts
