+++ packages/app/src/library/GenerateNominations/Prompts/SearchValidators.tsx
@@ -12,12 +12,13 @@ import { useState } from 'react'
 import { useTranslation } from 'react-i18next'
 import { ButtonPrimary } from 'ui-buttons'
 import { Checkbox } from 'ui-core/list'
-import type { PromptProps } from '../types'
+import type { PromptProps } from '../../../overlay/canvas/ManageNominations/types'
 
-export const FavoritesPrompt = ({ callback, nominations }: PromptProps) => {
+export const SearchValidators = ({ callback, nominations }: PromptProps) => {
   const { t } = useTranslation('modals')
   const { consts } = useApi()
   const { favoritesList } = useFavoriteValidators()
+
   const { maxNominations } = consts
 
   // Store the total number of selected favorites
@@ -37,30 +38,20 @@ export const FavoritesPrompt = ({ callback, nominations }: PromptProps) => {
 
   return (
     <>
-      <Title title={t('nominateFavorites')} />
+      <Title title={'Search Validators'} />
       <div className="padded">
-        {remaining.isLessThanOrEqualTo(0) ? (
-          <h4 className="subheading">
-            {t('moreFavoritesSurpassLimit', {
-              max: maxNominations.toString(),
-            })}
-          </h4>
-        ) : (
-          <h4 className="subheading">
-            {t('addUpToFavorites', { count: remaining.toNumber() })}.
-          </h4>
-        )}
+        <h4 className="subheading">Add validators to your nomination list.</h4>
 
         {favoritesList?.map((favorite: Validator, i) => {
           const inInitial = !!nominations.find(
             ({ address }) => address === favorite.address
           )
-          const isDisabled = selected.includes(favorite) || !canAdd || inInitial
+          const disabled = !canAdd || inInitial
 
           return (
             <PromptListItem
               key={`favorite_${i}`}
-              className={isDisabled && inInitial ? 'inactive' : undefined}
+              className={disabled && inInitial ? 'inactive' : undefined}
             >
               <Checkbox
                 checked={inInitial || selected.includes(favorite)}
@@ -76,6 +67,7 @@ export const FavoritesPrompt = ({ callback, nominations }: PromptProps) => {
             </PromptListItem>
           )
         })}
+
         <FooterWrapper>
           <ButtonPrimary
             text={t('addToNominations')}
